refactor: moved OutputFile to new module output_file
