+++ taskmaster/src/main.rs
@@ -1,6 +1,7 @@
 mod config;
 mod config_state;
 mod error;
+mod output_file;
 mod process_handler;
 mod tasks_manager;
