+++ taskmaster/src/process_handler/routine.rs
@@ -2,7 +2,8 @@ use super::KillCommandReceiver;
 use super::LogSender;
 use super::{Handle, NominativeStatus, Status, StatusSender, command};
 use crate::config::{AutoRestart, ProgramConfig};
-use crate::process_handler::{Log, LogType, OutputFile, Outputs};
+use crate::output_file::OutputFile;
+use crate::process_handler::{Log, LogType, Outputs};
 use libc::signal::kill;
 use libc::unistd::umask;
 use signal::Signal;
